refactor components and views
