--- src/components/ContractDeploy/index.ts
@@ -1,3 +1,0 @@
-import ContractDeploy from './ContractDeploy';
-
-export default ContractDeploy;
