+++ src/components/DebugDisplay/index.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect, useState } from 'react';
 import ReactDiffViewer, { DiffMethod } from 'react-diff-viewer';
 import './DebugDisplay.css';
-import { Button, ButtonType } from '../common/ui';
+import { Button, ButtonType } from '..';
 import { useForm } from 'react-hook-form';
 import { AppContext } from '../../appContext';
