+++ src/components/Deploy/DeployForm.tsx
@@ -1,8 +1,8 @@
 import React, { MutableRefObject, useContext, useEffect, useState, FormEvent, MouseEvent } from 'react';
 import { Controller, useForm } from 'react-hook-form';
-import { Button, ButtonType, TextArea } from 'components/common/ui';
+import { Button, ButtonType, TextArea } from '..';
 import { ABIDescription, ConstructorInput, ABIParameter } from 'types';
-import { abiHelpers } from '../common/lib';
+import { abiHelpers } from '../../lib';
 import { AppContext } from '../../appContext';
 
 interface IProps {
