--- src/types/miscTypes.ts
@@ -1,4 +1,0 @@
-export interface IFileSelectorItem {
-  value: string;
-  label: string;
-}
